[move] Move flock apps
