+++ eco-application/eco-application-example/src/main/kotlin/community/flock/eco/application/example/ApplicationData.kt
