+++ eco-application/eco-application-example/src/main/kotlin/community/flock/eco/application/example/authorities/ExampleAuthority.kt
