+++ eco-application/eco-application-multi_tenant/src/main/kotlin/community/flock/eco/application/multitenant/ApplicationConfiguration.kt
